api/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
python-dateutil==2.8.2
protobuf==3.20.1
Werkzeug==2.2.2
zipp==3.10.0
